refactor(mesh): rename sync_stream_messages -> sync_stream, move dispatch in
Two adjustments to the helper module landed in the prior commits of
this PR:

1. Rename `transport/sync_stream_messages.rs` -> `transport/sync_stream.rs`.
   The `_messages` suffix added no information; this file is the
   SyncStream RPC's message-level helpers, full stop.

2. Move `dispatch_stream_batch` from `transport/chunking.rs` into
   `transport/sync_stream.rs`. The function operates at the message
   layer (takes `StreamEntry`s unpacked from a `StreamBatch`, routes
   them through `MeshKV.notify_subscribers` and `ChunkAssembler`),
   not at the chunking primitive layer. Co-locating it with the
   outbound builders (`build_peer_stream_batches`, `wrap_stream_batch`,
   `build_heartbeat`) keeps the full SyncStream protocol round-trip
   in one file — every place a `StreamMessage` is constructed or
   applied is now in `sync_stream.rs`.

`chunking.rs` becomes pure data-shape primitives (`chunk_value`,
`build_stream_batches`, `next_generation`) with no `MeshKV` or
state coupling. Its module doc was updated to reflect the
narrower scope.
