+++ crates/mesh/src/tests/chunking_integration.rs
@@ -37,8 +37,9 @@ use crate::{
     kv::{MeshKV, StreamConfig, StreamRouting},
     service::gossip::StreamEntry,
     transport::{
-        chunking::{build_stream_batches, chunk_value, dispatch_stream_batch, next_generation},
+        chunking::{build_stream_batches, chunk_value, next_generation},
         limits::{DEFAULT_MAX_CHUNKS_PER_BATCH, MAX_STREAM_CHUNK_BYTES},
+        sync_stream::dispatch_stream_batch,
     },
 };
