+++ crates/mesh/src/transport/chunking.rs
@@ -1,13 +1,15 @@
-//! Sender-side stream chunking helpers and receiver-side dispatch.
+//! Stream chunking and batching primitives.
 //!
-//! Sender: [`chunk_value`] splits an oversized value into a sequence
-//! of `StreamEntry`s; [`build_stream_batches`] packs them into
-//! `StreamBatch` messages respecting a per-batch cap.
+//! [`chunk_value`] splits an oversized value into a sequence of
+//! `StreamEntry`s; [`build_stream_batches`] packs them into
+//! `StreamBatch` messages respecting per-batch count and byte caps.
+//! [`next_generation`] hands out monotonic tags so the receiver's
+//! [`ChunkAssembler`](crate::transport::chunk_assembler::ChunkAssembler)
+//! can group chunks of the same value.
 //!
-//! Receiver: [`dispatch_stream_batch`] routes inbound entries —
-//! single-chunk entries fire subscribers directly, multi-chunk entries
-//! go through the [`ChunkAssembler`](crate::transport::chunk_assembler::ChunkAssembler)
-//! and fire subscribers only on full reassembly.
+//! These are pure data-shape functions: no `MeshKV` state, no peer
+//! routing, no envelope construction. Higher-level composition lives
+//! in [`crate::transport::sync_stream`].
 
 use std::sync::{
     atomic::{AtomicU64, Ordering},
@@ -16,10 +18,7 @@ use std::sync::{
 
 use bytes::Bytes;
 
-use crate::{
-    kv::MeshKV,
-    service::gossip::{StreamBatch, StreamEntry},
-};
+use crate::service::gossip::{StreamBatch, StreamEntry};
 
 /// Monotonic generation counter for stream-batch values. Seeded from
 /// wall-clock nanos on first use so a process restart starts at a
@@ -89,52 +88,6 @@ pub fn chunk_value(
         });
     }
     entries
-}
-
-/// Receiver-side dispatch for `StreamBatch` entries. Single-chunk
-/// entries (`total_chunks == 1`) fire subscribers directly — no state
-/// in the chunk assembler. Multi-chunk entries route through the
-/// assembler and fire subscribers only on full reassembly.
-///
-/// Each chunk payload is detached from the decoded `StreamBatch`
-/// buffer via `Bytes::copy_from_slice` before it's stored or
-/// forwarded. Without this, the prost-generated `StreamEntry.data`
-/// is a `Bytes` view into the message frame, so a single pinned
-/// chunk (in the assembler or in a subscriber queue) would retain
-/// the entire batch allocation. That would let a peer packing many
-/// tiny entries into near-`MAX_MESSAGE_SIZE` batches defeat both
-/// `DEFAULT_MAX_ASSEMBLER_BYTES` and subscriber backpressure.
-///
-/// The chunk assembler scopes in-flight state by `peer_id` so
-/// concurrent chunked values from different senders under the same
-/// key don't collide.
-pub fn dispatch_stream_batch(
-    mesh_kv: &MeshKV,
-    peer_id: &str,
-    entries: impl IntoIterator<Item = StreamEntry>,
-) {
-    for entry in entries {
-        let data = Bytes::copy_from_slice(&entry.data);
-        if entry.total_chunks == 1 {
-            // A fresh single-chunk value supersedes any in-flight
-            // multi-chunk assembly for the same (peer, key); drop the
-            // stale fragments so they don't wait for GC.
-            mesh_kv.chunk_assembler().drop_pending(peer_id, &entry.key);
-            mesh_kv.notify_subscribers(&entry.key, Some(vec![data]));
-        } else {
-            let key = entry.key.clone();
-            if let Some(fragments) = mesh_kv.chunk_assembler().receive_chunk(
-                peer_id,
-                &key,
-                entry.generation,
-                entry.chunk_index,
-                entry.total_chunks,
-                data,
-            ) {
-                mesh_kv.notify_subscribers(&key, Some(fragments));
-            }
-        }
-    }
 }
 
 /// Pack `StreamEntry`s into one or more `StreamBatch`es, flushing
